changes in segmented buttons
