--- src/components/SegmentButton/index.ts
@@ -1,1 +1,0 @@
-;
