+++ src/components/SegmentedButtons/index.ts
@@ -1,0 +1,2 @@
+
+import {CustomSegmentedButtonsComponent } from './SegmentedButtons';
