+++ src/index.ts
@@ -1,2 +1,2 @@
 import './components/buttons/index' // import all buttons component here to the entry point for the npm package
-import "./components/SegmentButton/index"// import SegmentButton 
+import "./components/SegmentedButton/SegmentedButtons"// import SegmentButton
