+++ src/interface/SegmentedButtonsInterface/SegmentedButtonsInterface.ts
@@ -1,0 +1,9 @@
+ 
+ export interface SegmentedButtonsProps {
+  segments: { value: string; label: string }[];
+  selectedColor?: string;
+  unselectedColor?: string;
+  containerStyle?: any; 
+  theme?: any; 
+  onChange: (value: string) => void;
+}
